requirements.txt
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
asyncio==3.4.3
